0.9.8.1 исправлены баги с урезанием значения tot из double в int

+++ viewer/clogfilterpanel.h
@@ -19,7 +19,7 @@ public:
     ~ClogFilterPanel();
 
     void    setClusterRange(QVector<int> vector);
-    void    setTotRange(QVector<int> vector);
+    void    setTotRange(QVector<double> vector);
 
     int     getClusterBegin() const;
     int     getClusterEnd() const;
